refactor: flatten state module into sync
The extra nesting added no value.
