+++ crates/pathfinder/examples/verify_block_hashes.rs
@@ -2,7 +2,7 @@ use std::num::NonZeroU32;
 
 use anyhow::Context;
 use pathfinder_common::{BlockHash, BlockNumber, Chain, ChainId, StarknetVersion};
-use pathfinder_lib::state::block_hash::{verify_block_hash, VerifyResult};
+use pathfinder_lib::sync::block_hash::{verify_block_hash, VerifyResult};
 use pathfinder_storage::{JournalMode, Storage};
 use stark_hash::Felt;
 use starknet_gateway_types::reply::{Block, Status};
