+++ crates/pathfinder/src/lib.rs
@@ -1,7 +1,7 @@
 #![deny(rust_2018_idioms)]
 
 pub mod monitoring;
-pub mod state;
+pub mod sync;
 
 #[cfg(feature = "p2p")]
 pub mod p2p_network;
