--- crates/pathfinder/src/state.rs
@@ -1,4 +1,0 @@
-pub mod block_hash;
-mod sync;
-
-pub use sync::{l1, l2, sync, SyncContext};
