+++ crates/pathfinder/src/sync.rs
@@ -1,0 +1,8 @@
+pub mod block_hash;
+mod class;
+pub mod l1;
+pub mod l2;
+mod pending;
+mod sync;
+
+pub use sync::{sync, SyncContext};
