+++ crates/pathfinder/src/sync/class.rs
