+++ crates/pathfinder/src/sync/l1.rs
@@ -6,7 +6,7 @@ use pathfinder_retry::Retry;
 use primitive_types::H160;
 use tokio::sync::mpsc;
 
-use crate::state::sync::SyncEvent;
+use crate::sync::sync::SyncEvent;
 
 #[derive(Clone)]
 pub struct L1SyncContext<EthereumClient> {
